VocalGuard/requirements.txt
# -------- Flask & Server --------
Flask==2.3.2
Flask-Cors==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0

# -------- TensorFlow Lite & ML --------
tensorflow-cpu==2.17.0
numpy==1.26.4
joblib==1.3.2

# -------- Audio Processing --------
numba==0.58.1
llvmlite==0.41.1
librosa==0.10.2.post1
soundfile==0.12.1
pydub
pandas